% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{LIR.XIC.pair}
\alias{LIR.XIC.pair}
\alias{LIR.AIC.pair}
\alias{LIR.BIC.pair}
\alias{LIR.QAIC.pair}
\title{Other model selection criterion(AIC/BIC/QAIC)

Not recommended for composite likelihood model.}
\usage{
LIR.AIC.pair(theta, model, ni, nj, m, tau, ..., MCLE = FALSE)

LIR.BIC.pair(theta, model, ni, nj, m, tau, T, ..., MCLE = FALSE)

LIR.QAIC.pair(theta, model, ni, nj, m, tau, ..., MCLE = FALSE)
}
\arguments{
\item{theta}{Parameter to calculate \eqn{\hat{R_{\tau}}}. If param `MCLE` is TRUE,
this is the MCLE, otherwise the initial value for optimizer. Theta is required
because it is the only parameter that tells the dimension of estimating variable.}

\item{model}{Model to calculate \eqn{\hat{R_{\tau}}}. Should be a function
that takes theta, tau_i(and other arguments if necessary) and return a
numeric number}

\item{ni}{Number of individuals at each observation}

\item{nj}{Number of individuals at each lagged observation}

\item{m}{Vector of lagged identification for all observation pair}

\item{tau}{Vector of time interval of lagged identification}

\item{...}{Additional parameters passed to model, or upper/lower bound and optimize
argument when calculating MCLE.}

\item{MCLE}{If TRUE param `theta` is the MCLE, otherwise `theta` will be used as
the initial value for optimizer. Boolean, default FALSE.}

\item{T}{Total study time. Required for BIC.}
}
\value{
score
}
\description{
`LIR.AIC.pair` is for AIC with pairwise data.
`LIR.BIC.pair` is for BIC with pairwise data.
`LIR.QAIC.pair` is for QIC with pairwise data.
\deqn{AIC(\hat{\theta}_{LIR})=-2\sum_{t_i \in \tau_0}\sum_{\tau \in M}\{cl_{ij}(\hat{\theta}_{LIR})\}+2k}
\deqn{BIC(\hat{\theta}_{LIR})=-2\sum_{t_i \in \tau_0}\sum_{\tau \in M}\{cl_{ij}(\hat{\theta}_{LIR})\}+k ln(T)}
\deqn{QAIC(\hat{\theta}_{LIR})=-2\sum_{t_i \in \tau_0}\sum_{\tau \in M}\{cl_{ij}(\hat{\theta}_{LIR})\}/\hat{c}+2k}
Here \eqn{k} is `length(theta)`, T is total study time, \eqn{\hat{c}} is the
variance inflation factor estimated from the ratio of the goodness-of-fit
\eqn{\chi^2}-statistic to its degrees of freedom.
}
