% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LIR.chat}
\alias{LIR.chat}
\title{Function for VIF(Variance Inflation Factor) in QAIC}
\usage{
LIR.chat(data, tp, k, mtau = -1)
}
\arguments{
\item{data}{Observation matrix}

\item{tp}{List-like observation time(1d vector)}

\item{k}{the number of variable in the most general model}

\item{mtau}{The maximum allowable lag time. If a lagged pair has time \eqn{\tau}
greater than `mtau`, it will not be used to calculate composite likelihood.
If `mtau` is less than zero, all pairs will be used. Default -1.0.}
}
\description{
\deqn{\hat{c}=\chi^2/df=\sum_{i,\tau}\frac{(m_{t_i,t_i+\tau}-E(m_{t_i,t_i+\tau}))^2}{E(m_{t_i,t_i+\tau})}/(unique(\tau)-k-1)}
}
\note{
This function has memory complexity of O(N^2) and may suffer from out-of-memory
error if observation number is very large.
}
