% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{LIR.CLICa/b}
\alias{LIR.CLICa/b}
\alias{LIR.CLICa}
\alias{LIR.CLICb}
\title{Composite likelihood Information Criterion}
\usage{
LIR.CLICa(
  theta,
  model,
  grad,
  hessian,
  data,
  t,
  ...,
  MCLE = FALSE,
  B = 500,
  cl = NULL,
  ncores = -1
)

LIR.CLICb(
  theta,
  model,
  grad,
  hessian,
  data,
  t,
  T,
  ...,
  MCLE = FALSE,
  B = 500,
  cl = NULL,
  ncores = -1
)
}
\arguments{
\item{theta}{Parameter to calculate \eqn{\hat{R_{\tau}}}. If param `MCLE` is TRUE,
this is the MCLE, otherwise the initial value for optimizer. Theta is required
because it is the only parameter that tells the dimension of estimating variable.}

\item{model}{Model to calculate \eqn{\hat{R_{\tau}}}. Should be a function
that takes theta, tau_i(and other arguments if necessary) and return a
numeric number}

\item{grad}{Gradient of the model. Should be a function that takes theta,
tau_i(and other arguments if necessary) and return a 1D
vector with same length as theta}

\item{hessian}{Hessian of the model. Should be a function that takes theta,
tau_i(and other arguments if necessary) and return a square symmetric matrix
with length(theta) rows and columns}

\item{data}{Observation matrix}

\item{t}{List-like observation time(1d vector)}

\item{...}{Additional parameters passed to model, or upper/lower bound and optimize
argument when calculating MCLE.}

\item{MCLE}{If TRUE param `theta` is the MCLE, otherwise `theta` will be used as
the initial value for optimizer. Boolean, default FALSE.}

\item{B}{Bootstrap's repeat sampling times}

\item{cl}{Cluster to use, Default NULL. If NULL, a new cluster will be created by @seealso [makeCluster()]}

\item{ncores}{Number of processors to use. Default -1(which means all available cores).
This argument will be suppressed if cl is not NULL.}

\item{T}{Total study time. Required for CLICb.}
}
\value{
CLIC score
}
\description{
`LIR.CLICa` is CLICa criterion and `LIR.CLICb` is CLICb criterion.
\deqn{CLIC_a(\hat{\theta}_{LIR})=-2\sum_{t_i \in \tau_0}\sum_{\tau \in M}\{cl_{ij}(\hat{\theta}_{LIR})\}+2tr\{J_T(\theta_{LIR})H_T(\theta_{LIR})^{-1}\}}
\deqn{CLIC_b(\hat{\theta}_{LIR})=-2\sum_{t_i \in \tau_0}\sum_{\tau \in M}\{cl_{ij}(\hat{\theta}_{LIR})\}+ln(T)tr\{J_T(\theta_{LIR})H_T(\theta_{LIR})^{-1}\}}
\eqn{tr\{J_T(\theta_{LIR})H_T(\theta_{LIR})^{-1}\}} is difficult to calculate
thus replaced by \eqn{tr\{H_T(\theta_{LIR})var(\hat{\theta}_{LIR})\}}, which is more
robust. Bootstrap of individual is applied to estimate covariance matrix. Re-sample
procedure is paralleled.
}
