% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{LIR.grad.A/B/C}
\alias{LIR.grad.A/B/C}
\alias{LIR.grad.A}
\alias{LIR.grad.B}
\alias{LIR.grad.C}
\title{Gradient of model A/B/C}
\usage{
LIR.grad.A(theta, tau, ...)

LIR.grad.B(theta, tau, ...)

LIR.grad.C(theta, tau, ...)
}
\arguments{
\item{theta}{Parameter to estimate.}

\item{tau}{Lagged time \eqn{\tau}.}
}
\value{
vector
}
\description{
Gradient of model A/B/C
}
