% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcle.R
\name{LIR.CLhessian.pair}
\alias{LIR.CLhessian.pair}
\title{Hessian of composite likelihood for pairwise input}
\usage{
LIR.CLhessian.pair(theta, model, grad, hessian, ni, nj, m, tau, ...)
}
\arguments{
\item{theta}{Parameter to calculate \eqn{\hat{R_{\tau}}}}

\item{model}{Model to calculate \eqn{\hat{R_{\tau}}}. Should be a function
that takes theta, tau_i(and other arguments if necessary) and return a
numeric number}

\item{grad}{Gradient of the model. Should be a function that takes theta,
tau_i(and other arguments if necessary) and return a 1D
vector with same length as theta}

\item{hessian}{Hessian of the model. Should be a function that takes theta,
tau_i(and other arguments if necessary) and return a square symmetric matrix
with length(theta) rows and columns}

\item{ni}{of individuals of the former observation at each lagged time pair.}

\item{nj}{Number of individuals of the latter observation at each lagged time pair.}

\item{m}{vector of total lagged identification number for all lagged time pair}

\item{tau}{vector of time interval of lagged identification}

\item{...}{extra arguments to be passed to model to calculate \eqn{\hat{R_{\tau}}}}
}
\value{
A square symmetric matrix with length(theta) rows and columns.
}
\description{
Hessian of composite likelihood for pairwise input
}
