% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{LIR.model.A/B/C}
\alias{LIR.model.A/B/C}
\alias{LIR.model.A}
\alias{LIR.model.B}
\alias{LIR.model.C}
\title{LIR model A/B/C}
\usage{
LIR.model.A(theta, tau, ...)

LIR.model.B(theta, tau, ...)

LIR.model.C(theta, tau, ...)
}
\arguments{
\item{theta}{Parameter to estimate.}

\item{tau}{Lagged time \eqn{\tau}.}
}
\value{
\eqn{R_{\tau}}
}
\description{
A: \eqn{R_{\tau}=\alpha, \theta=\alpha=1/N}

B: \eqn{R_{\tau}=\alpha e^{-\beta\tau}, \theta=c(\alpha, \beta)=c(1/N, \lambda)}

C: \eqn{R_{\tau}=\gamma e^{-\beta\tau}+\alpha, \theta=c(\gamma, \beta, \alpha)=c(\frac{\lambda}{(\lambda+\mu)N}, \lambda+\mu, \frac{\mu}{(\lambda+\mu)N})}
}
\details{
In model A, it is assumed that no migration occurs and population N remains constant.

In model B, move-in rate equals to move-out rate so the population remains constant.
Note that migration in this model is permanent, so animals previously move out(or dead)
will not return.

In model C, animals in the study area move out with probability of \eqn{\lambda} per
unit time and move in with probability of \eqn{\mu} per unit time. The population of the
whole area T is assumed to be constant. If \eqn{\lambda = \frac{\mu (Z-N)}{N}}, the
population within the study has an expectation of N, otherwise a warning will
be raised.
}
