% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcle.R
\name{move.CL}
\alias{move.CL}
\title{Composite likelihood function for estimating LIR(Lagged Identification Rate)
the notation of the estimation is \eqn{\hat{R_{tau}}}}
\usage{
move.CL(theta, model, n, N, data, ...)
}
\arguments{
\item{theta}{Parameter to calculate R_tau}

\item{model}{Model to calculate R_tau}

\item{n}{number of individuals at each observation}

\item{N}{population}
}
\value{
Likelihood score(numeric). Inf if theta out of \link{0,1}
}
\description{
Composite likelihood function for estimating LIR(Lagged Identification Rate)
the notation of the estimation is \eqn{\hat{R_{tau}}}
}
