% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcle.R
\name{move.model.A}
\alias{move.model.A}
\title{LIR model A. The population is constant and no migrations occur.}
\usage{
move.model.A(theta)
}
\arguments{
\item{theta}{here theta is the estimated LIR(R_tau) directly}
}
\value{
theta
}
\description{
LIR model A. The population is constant and no migrations occur.
}
