% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcle.R
\name{move.model.B}
\alias{move.model.B}
\title{LIR model B.}
\usage{
move.model.B(theta, tau)
}
\arguments{
\item{theta}{(\eqn{\alpha, \beta})}

\item{tau}{List of time interval of lagged identification}
}
\value{
List of estimated R_tau
}
\description{
LIR model B.
}
