% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{move.pairwise}
\alias{move.pairwise}
\title{Calculate lagged identification of each observation pair.
Number of observation is guessed by the column number of input matrix}
\usage{
move.pairwise(data, t = NULL, Time = NULL, tau = TRUE)
}
\arguments{
\item{data}{Matrix of identification with row number equal to population
and column number equal to observation.}

\item{t}{List like time of observation. Default NULL, must given if tau is true}

\item{tau}{whether return tau}
}
\value{
List of lagged identification number for each observation pair.
}
\description{
Calculate lagged identification of each observation pair.
Number of observation is guessed by the column number of input matrix
}
