% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{move.simulate.A}
\alias{move.simulate.A}
\title{Simulate animal movement identification data with model A
Population is constant and no migration.}
\usage{
move.simulate.A(N, n, t, seed = NULL)
}
\arguments{
\item{N}{population}

\item{n}{number of observed individual at each moment. List or integer}

\item{t}{Time of each observation if t is a list or array, the number
of observation if t is an integer.}

\item{seed}{random seed}
}
\value{
0-1 identification matrix of N rows and length_t(or t) columns
}
\description{
Simulate animal movement identification data with model A
Population is constant and no migration.
}
