% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{move.simulate.B}
\alias{move.simulate.B}
\title{Simulate animal movement identification data with model B}
\usage{
move.simulate.B(N, T, n, t, lambda, seed = NULL)
}
\arguments{
\item{N}{Population with the study area}

\item{n}{Number of identification in each observation}

\item{t}{Time of each observation if t is a list or array, the number
of observation if t is an integer.}

\item{lambda}{migration rate}

\item{seed}{random seed}
}
\value{
matrix of N rows and length_t(or t) columns
}
\description{
Simulate animal movement identification data with model B
}
