% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{move.simulate.C}
\alias{move.simulate.C}
\title{Simulate animal movement identification data with model C}
\usage{
move.simulate.C(Z, N, T, n, t, lambda, mu, seed = NULL)
}
\arguments{
\item{Z}{Total population of the whole area}

\item{N}{Initial population with the study area}

\item{n}{Number of identification in each observation}

\item{t}{Time of each observation if t is a list or array, the number
of observation if t is an integer.}

\item{lambda}{move out rate}

\item{mu}{move in rate}

\item{seed}{random seed}
}
\value{
matrix of Z rows and length(t) columns
}
\description{
Simulate animal movement identification data with model C
}
