% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{LIR.XIC/LIR.XIC.pair}
\alias{LIR.XIC/LIR.XIC.pair}
\alias{LIR.AIC.pair}
\alias{LIR.AIC}
\alias{LIR.BIC}
\alias{LIR.QAIC.pair}
\alias{LIR.QAIC}
\title{Other model selection criterion(AIC/BIC/QAIC)

Not recommended for composite likelihood model.}
\usage{
LIR.AIC.pair(theta, model, ni, nj, m, tau, ..., MCLE = FALSE)

LIR.AIC(
  theta,
  model,
  data,
  tp,
  ...,
  model_args = list(),
  mtau = Inf,
  MCLE = FALSE
)

LIR.BIC(
  theta,
  model,
  data,
  tp,
  ...,
  model_args = list(),
  mtau = Inf,
  MCLE = FALSE
)

LIR.QAIC.pair(c, theta, model, ni, nj, m, tau, ..., MCLE = FALSE)

LIR.QAIC(
  c,
  theta,
  model,
  data,
  tp,
  ...,
  model_args = list(),
  mtau = Inf,
  MCLE = FALSE
)
}
\arguments{
\item{theta}{Parameter to calculate \eqn{\hat{R_{\tau}}}. If param `MCLE` is TRUE,
this is the MCLE, otherwise the initial value for optimizer. Theta is required
because it is the only parameter that tells the dimension of estimating variable.}

\item{model}{Model to calculate \eqn{\hat{R_{\tau}}}. Should be a function
that takes theta, tau_i(and other arguments if necessary) and return a
numeric number}

\item{ni}{Number of individuals at each observation}

\item{nj}{Number of individuals at each lagged observation}

\item{m}{Vector of lagged identification for all observation pair}

\item{tau}{Vector of time interval of lagged identification}

\item{...}{Additional optimizer parameters
argument when calculating MCLE.}

\item{MCLE}{If TRUE then `theta` is the MCLE, otherwise `theta` will
be used as the initial value for optimizer. Default FALSE.}

\item{data}{Observation matrix}

\item{tp}{List-like observation time(1d vector)}

\item{model_args}{Additional parameters passed to model}

\item{mtau}{The maximum allowable lag time. If a lagged pair has time \eqn{\tau}
greater than `mtau`, it will not be used to calculate composite likelihood.
If `mtau` is less than zero, all pairs will be used. Default -1.0.}

\item{c}{VIF(Variance Inflation Factor) for QAIC}
}
\value{
score
}
\description{
`LIR.AIC` and `LIR.AIC.pair`(deprecated) is for AIC.
`LIR.BIC` is for BIC.
`LIR.QAIC` and `LIR.QAIC.pair`(deprecated) is for QAIC.
\deqn{AIC(\hat{\theta}_{LIR})=-2\sum_{t_i \in \tau_0}\sum_{\tau \in M}\{cl_{ij}(\hat{\theta}_{LIR})\}+2k}
\deqn{QAIC(\hat{\theta}_{LIR})=-2\sum_{t_i \in \tau_0}\sum_{\tau \in M}\{cl_{ij}(\hat{\theta}_{LIR})\}/\hat{c}+2k}
Here \eqn{k} is `length(theta)`, T is number of observation time,
\eqn{\hat{c}} is the variance inflation factor estimated from the
ratio of the goodness-of-fit \eqn{\chi^2}-statistic to its degrees of freedom.
}
\references{
Hal Whitehead (2007) Selection of Models of Lagged Identification Rates
and Lagged Association Rates Using AIC and QAIC, Communications in Statistics - Simulation and
Computation, 36:6, 1233-1246, DOI: 10.1080/03610910701569531
}
\seealso{
chat
}
