% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/mcle.R
\name{LIR.CL}
\alias{LIR.CL}
\alias{LIR.CL.pair}
\title{Composite likelihood for LIR}
\usage{
LIR.CL(theta)

LIR.CL.pair(theta, model, ni, nj, m, tau, ...)
}
\arguments{
\item{theta}{Parameter to calculate \eqn{\hat{R_{\tau}}}}

\item{model}{Model to calculate \eqn{\hat{R_{\tau}}}. Should be a function
that takes `theta`, `tau_i`(and other arguments if necessary) and return a
numeric number}

\item{ni}{Number of individuals of the former observation at each lagged time pair.}

\item{nj}{Number of individuals of the latter observation at each lagged time pair.}

\item{m}{vector of total lagged identification number for all lagged time pair}

\item{tau}{vector of time interval of lagged identification}

\item{...}{extra arguments to be passed to model to calculate \eqn{\hat{R_{\tau}}}}

\item{data}{Observation matrix}

\item{t}{List-like observation time(1d vector)}
}
\value{
minus Likelihood score(numeric). Inf if theta out of [0,1]
}
\description{
This CL function requires a sample format of pairwise lagged identification,
which is a list-like data containing lagged identification of of all pairs of
observation. This function is recommended when total observation time is very small and
pairwise lagged identification is easy to pre-calculate through `LIR.pairwise()`.
The notation of the estimation is \eqn{\hat{R_{\tau}}}.
}
