% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/mcle.R
\name{LIR.CLgrad}
\alias{LIR.CLgrad}
\alias{LIR.CLgrad.pair}
\title{Gradient of composite likelihood}
\usage{
LIR.CLgrad(theta, model, grad, data, tp, model_args, mtau = -1)

LIR.CLgrad.pair(theta, model, grad, ni, nj, m, tau, model_args)
}
\arguments{
\item{theta}{Parameter to calculate \eqn{\hat{R_{\tau}}}}

\item{model}{Model to calculate \eqn{\hat{R_{\tau}}}. Should be a function
that takes theta, tau_i(and other arguments if necessary) and return a
numeric number}

\item{grad}{Gradient of the model. Should be a function that takes theta,
tau_i(and other arguments if necessary) and return a 1D
vector with same length as theta}

\item{data}{Observation matrix}

\item{tp}{List-like observation time(1d vector)}

\item{model_args}{extra arguments to be passed to model to calculate
\eqn{\hat{R_{\tau}}}. If there are no additional parameters, pass a empty
list or `NULL`.}

\item{mtau}{The maximum allowable lag time. If a lagged pair has time \eqn{\tau}
greater than `mtau`, it will not be used to calculate composite likelihood.
If `mtau` is less than zero, all pairs will be used. Default -1.0.}

\item{ni}{of individuals of the former observation at each lagged time pair.}

\item{nj}{Number of individuals of the latter observation at each lagged time pair.}

\item{m}{vector of total lagged identification number for all lagged time pair}

\item{tau}{vector of time interval of lagged identification}
}
\value{
A 1D vector with same length as theta
}
\description{
Gradient of composite likelihood
}
