% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{LIR.hessian.A/B/C}
\alias{LIR.hessian.A/B/C}
\alias{LIR.hessian.A}
\alias{LIR.hessian.B}
\alias{LIR.hessian.C}
\title{Hessian matrix of model A/B/C}
\usage{
LIR.hessian.A(theta, tau, ...)

LIR.hessian.B(theta, tau, ...)

LIR.hessian.C(theta, tau, ...)
}
\arguments{
\item{theta}{Parameter to estimate.}

\item{tau}{Lagged time \eqn{\tau}.}
}
\value{
A symmetric matrix.
}
\description{
Hessian matrix of model A/B/C
}
