% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{LIR.modelSelect}
\alias{LIR.modelSelect}
\title{Model Selection for lagged identification rate model.}
\usage{
LIR.modelSelect(
  data,
  tp,
  theta_list = list(rep(0.1, 1), rep(0.001, 2), rep(0.001, 3)),
  model_list = list("A", "B", "C"),
  criterion = "CLICa",
  ...,
  model_args = list(),
  mtau = -1,
  MCLE = FALSE,
  B = 500,
  cl = NULL,
  ncores = -1,
  verbose = TRUE
)
}
\arguments{
\item{data}{Observation matrix}

\item{tp}{List-like observation time(1d vector)}

\item{theta_list}{List of Parameter to calculate \eqn{\hat{R_{\tau}}}
correspond to `model_list`. If param `MCLE` is TRUE,
this is the MCLE, otherwise the initial value for optimizer. Theta is required
because it is the only parameter that tells the dimension of estimating variable.}

\item{model_list}{List of candidate models, each element of this list should
be 'A', 'B', 'C', a function, or a vector of three function.
If it is 'A'/'B'/'C' then built-in model A/B/C will be used. If the criterion
is CLICa/b, then user-defined model must be a vector of three function, function
to calculate LIR \eqn{R_{\tau}}, gradient of \eqn{R_{\tau}} and Hessian matrix
of \eqn{R_{\tau}}. If the criterion is AIC/BIC/QAIC, gradients and Hessian matrix
is not needed and user-defined model could be a single function. Default list('A', 'B', 'C').}

\item{criterion}{Model selection criterion. Should be one of 'AIC', 'BIC',
'QAIC', 'CLICa' and 'CLICb'.}

\item{...}{Additional optimizer parameters for MCLE.}

\item{model_args}{Additional parameters passed to model}

\item{mtau}{The maximum allowable lag time. If a lagged pair has time \eqn{\tau}
greater than `mtau`, it will not be used to calculate composite likelihood.
If `mtau` is less than zero, all pairs will be used. Default -1.0.}

\item{MCLE}{If TRUE then `theta` is the MCLE, otherwise `theta` will be used as
the initial value for optimizer. Boolean, default FALSE.}

\item{B}{Bootstrap's repeat sampling times}

\item{cl}{Cluster to use for CLICa/b, Default NULL. If NULL, a new cluster
will be created by @seealso [makeCluster()] and closed after function finished.
If not NULL, the function will use the given cluster and will not close it.}

\item{ncores}{Number of processors to use. Default -1 (which means all
available cores). This argument will be suppressed if cl is not NULL.}

\item{verbose}{Whether return verbose output. Default TRUE.}
}
\value{
If `verbose` is FALSE return the score list of each model only. Otherwise
return a list.
\describe{
  \item{best}{Best model name}
  \item{best_model}{Best model function}
  \item{theta}{MCLE param to the best model}
  \item{score}{List of score to each model in `model_list`}
  \item{model_list}{Candidate model list}
  \item{criterion}{Criterion of model selection, characters}
}
}
\description{
List of model should be provided and a specific criterion among
AIC/QAIC/CLICa/CLICb should be assigned. This function will calculate
score of each model with the given criterion and select the model with highest
score as the best one.
}
\note{
`theta_list` must be correspond to `model_list`. Compactbility between
theta and LIR models will not be examined.
}
\examples{
# Example of model selection
# Set observation time
tp <- c(1:5, 51:55, 101:105, 501:505, 601:605)
# Generate observation matrix with model C
data <- move.simulate.C(300, 100, 605, 40, 0.08, 0.04)
# Model Selection
# Initial value of theta and boundary(when length(theta)==1) are very important
ms <- LIR.modelSelect(data, tp, B=4, MCLE=FALSE, lower=0.0005, upper=0.09)
ms$score
# [1] 1.800000e+13 1.099621e+04 1.074243e+04
ms$best
# [1] "C"
ms$theta
# [1] 0.007119308 0.212260232 0.003582160

}
