% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{LIR.pairwise}
\alias{LIR.pairwise}
\title{Calculate lagged identification of each observation pair.}
\usage{
LIR.pairwise(data, tp = NULL, require_tau = TRUE, require_n = TRUE)
}
\arguments{
\item{data}{Matrix of identification with row number equal to population
and column number equal to observation.}

\item{tp}{List like time of observation. Default NULL, must given if tau is true}

\item{require_tau}{If true, return tau}

\item{require_n}{If true, return n}
}
\value{
List of lagged identification number for each observation pair.
}
\description{
Calculate lagged identification of each observation pair. Number of observation is guessed by the column number of input matrix
}
